Extract audio-processor metrics
