+++ crates/apps/looper/looper-processor/src/audio/multi_track_looper/metrics/mod.rs
@@ -1,2 +1,2 @@
-pub mod audio_processor_metrics;
+pub use augmented_audio_metrics::audio_processor_metrics;
 pub mod audio_thread_logger;
