+++ crates/augmented/ops/augmented-metrics/src/lib.rs
@@ -1,5 +1,8 @@
 use std::time::Instant;
 
+pub mod audio_processor_metrics;
+
+/// Log duration of a function
 pub fn time<T>(label: &str, body: impl FnOnce() -> T) -> T {
     let start = Instant::now();
     let result = body();
